Crate webvr_traits

Re-exports

pub use rust_webvr_api as webvr;

Modules

webvr_traits

Structs

MockVRInit

MockVRView

WebVRDisplayCapabilities
describes the capabilities of a VRDisplay. These are expected to be static per-device/per-user.

WebVRDisplayData

WebVREyeParameters
The VREyeParameters interface represents all the information required to correctly render a scene for a given eye.

WebVRFieldOfView
The VRFieldOfView interface represents a field of view, as given by 4 degrees describing the view from a center point.

WebVRFrameData
Represents all the information needed to render a single frame of a VR scene

WebVRFutureFrameData

WebVRGamepadButton

WebVRGamepadData

WebVRGamepadState

WebVRLayer
Data provided to a VRDisplay and presented in the HMD.

WebVRPose
The VRPose struct represents a sensor’s state at a given timestamp.

WebVRPoseInformation

WebVRStageParameters
The VRStageParameters interface represents the values describing the stage/play area for displays that support room-scale experiences.

Enums

MockVRControlMsg

WebVRDisplayEvent

WebVRDisplayEventReason

WebVREvent

WebVREye
VREye

WebVRGamepadEvent

WebVRGamepadHand

WebVRMsg

Constants

_IMPL_DESERIALIZE_FOR_WebVRPoseInformation

_IMPL_SERIALIZE_FOR_WebVRPoseInformation

Traits

WebVRMainThreadHeartbeat
Each VR service may have some code which needs to run on the main thread, for example window creation on MacOS is only supported on the main thread. Implementations of this trait will usually be neither Sync nor Send.

Type Definitions

WebVRResult